Reorganizing utilities and fixing tests
