+++ dev/testing/microsoft-agents-testing/microsoft_agents/testing/utils/send.py
@@ -11,10 +11,8 @@ from microsoft_agents.activity import Activity
 from microsoft_agents.testing.core import (
     Exchange,
     ExternalScenario,
-    Scenario,
 )
 from microsoft_agents.testing.core.utils import activities_from_ex
-from .scenario_registry import scenario_registry
 
 def _create_activity(payload: str | dict | Activity) -> Activity:
     """Create an Activity from various payload types.
@@ -83,23 +81,3 @@ async def send(
     """
     exchanges = await ex_send(payload, url, listen_duration)
     return activities_from_ex(exchanges)
-
-def resolve_scenario(scenario_or_str: Scenario | str ) -> Scenario:
-    """Resolve a scenario from a Scenario instance or a registered name.
-    
-    If a string is provided, looks up the scenario in the registry.
-    
-    :param scenario_or_str: A Scenario instance or a string key for lookup.
-    :return: The resolved Scenario instance.
-    :raises ValueError: If the string key is not found in the registry.
-    """
-    if isinstance(scenario_or_str, Scenario):
-        return scenario_or_str
-    elif isinstance(scenario_or_str, str):
-        if scenario_or_str.startswith("http://") or scenario_or_str.startswith("https://"):
-            # If it's a URL, create an ExternalScenario on the fly
-            return ExternalScenario(scenario_or_str)
-        else:
-            return scenario_registry.get(scenario_or_str)
-    else:
-        raise TypeError("Input must be a Scenario instance or a string key.")
